+++ auction-server/src/main/java/com/group13/auction/service/iservice/IWalletService.java
@@ -24,7 +24,7 @@ public interface IWalletService {
      * @param amount số tiền (phải > 0)
      * @throws IllegalArgumentException nếu amount <= 0
      */
-    void deposit(NormalUser user, double amount);
+    void deposit(NormalUser user, long amount);
 
     /**
      * Rút tiền từ tài khoản NormalUser.
@@ -37,7 +37,7 @@ public interface IWalletService {
      * @throws IllegalArgumentException nếu amount <= 0 hoặc vượt số dư khả dụng
      * @throws IllegalStateException nếu tài khoản không đủ điều kiện
      */
-    void withdraw(NormalUser user, double amount);
+    void withdraw(NormalUser user, long amount);
 
     /**
      * Khóa cọc khi bidder tham gia phiên đấu giá (joinAuction).
@@ -50,7 +50,7 @@ public interface IWalletService {
      * @throws com.group13.auction.exception.AuctionBusinessException
      *         nếu số dư không đủ để khóa cọc
      */
-    void lockDeposit(NormalUser bidder, double depositAmount, String auctionId);
+    void lockDeposit(NormalUser bidder, long depositAmount, String auctionId);
 
     /**
      * Hoàn cọc cho bidder không thắng khi phiên kết thúc.
@@ -60,7 +60,7 @@ public interface IWalletService {
      * @param depositAmount số tiền cọc được hoàn
      * @param auctionId     id phiên
      */
-    void unlockDeposit(NormalUser bidder, double depositAmount, String auctionId);
+    void unlockDeposit(NormalUser bidder, long depositAmount, String auctionId);
 
     /**
      * Tịch thu cọc của winner không thanh toán đúng hạn.
@@ -71,7 +71,7 @@ public interface IWalletService {
      * @param depositAmount số tiền cọc bị tịch thu
      * @param auctionId     id phiên
      */
-    void forfeitDeposit(NormalUser winner, double depositAmount, String auctionId);
+    void forfeitDeposit(NormalUser winner, long depositAmount, String auctionId);
 
     /**
      * Thực hiện toàn bộ luồng giao dịch thanh toán sau khi phiên FINISHED.
@@ -87,7 +87,7 @@ public interface IWalletService {
      * @throws com.group13.auction.exception.PaymentException nếu winner không đủ số dư
      */
     void executePaymentTransaction(NormalUser winner, NormalUser seller,
-                                   double finalPrice, double depositPaid, String auctionId);
+                                   long finalPrice, long depositPaid, String auctionId);
 
     /**
      * Hoàn tiền 100% cho winner khi seller vi phạm chất lượng hàng hóa.
@@ -99,7 +99,7 @@ public interface IWalletService {
      * @param auctionId  id phiên
      */
     void executeRefundToWinner(NormalUser winner, NormalUser seller,
-                               double finalPrice, String auctionId);
+                               long finalPrice, String auctionId);
 
     /**
      * Thực hiện giao dịch Second Chance Offer khi runner-up chấp nhận mua.
